enhc Add full width, featured background image
This is a "featured" project image that spans from top of the page
til the next section.

Uses a new background image class and user applies a background url
along with a height inline.

Not 100% confident in my choice to use z-indexing on this solution, but
am also not 100% sure on how this bg image and other will be used
across the site.

Other changes:
- Setup z-index variables to avoid collisions.