enhc Rectangular progress bar
Creates a rectangular progress bar which allows lines and content to
highlist specific parts of it. This was implemented using a relative
positioning interface.

This has following strengths and weaknesses:
- To position elements above and below the bar, only simple markup
is required and not complex css classes
- Because the lines are not being set absolutely, the "left" percentage
values are not exact in where they show up. So 100% will show up in the
incorrect position, so dev would have to fine tune.

TODO:
- implement this on project page.