ehnc Backend for sorter
I'd like some advice on the following:
- How I passed a []Volunteer object down. I think theres a better way
to do this, and some nice python way of copy those properties over
- My naming of the new routes. I referred to REST principals of creating
a new resource for this as one doens't really exist for it. Is "all" ok?

TODO:
- I found an issue where the multiple <object> tags are not always using
the cache to get the svg image. This is causing performance problems and
has to be fixed.