Added testing using tape and testling to run
Using tapes to write and testling to run the unit tests in a headless
browser.

Eventually can use testling ci for browser tests, but this is not
a requirement.

Also am using sinon js for mocking.

Setup testing for update_percent file as example. This is full DOM
and ajax testing.

Other changes:
* Added a grunt task for timing the whole build. I think its important
that it stays under a certain time.