ehnc Project filter full html setup
Changed the way we structure the HTML in the template. Before we were
just dumping all the projects at the end. Now I'm adding the projects
under their place.

Reasons:
We were making the assumption that less html and moving things around
the DOM with JS would mean better performance. I realized this might
be an incorrect assumption, considering the amount of DOM manipulations
so I decided to setup this way because it's the easier solution to
start out with. I can branch and create code the other way later to
compare.

* This means there will be duplicate html in the template.
* This means the JS will not have to link the projects to their parent
issue, country and volunteer, meaning less JS.

TODO:
* I haven't formatted country yet as @cm said to wait til we have to
do a full database reset.