Changed html structure and js code for project sorter
Changed htlm structure to not be table as discussed.

This is not completely to design specs yet. Currently, its missing the
first "page" on the murally. This means that the campaign fund and all
associated projects are there and are filterable with the nav, but not
the single line titles of the countries and issues.

I also haven't done a separate rendering for the volunteer vs country
vs issues project yet, meaning they all have the same information and
layout right now.