Donation form top part
Renders a version of the top of the donation form for both projects
and funds.

I was messing around with the best way to pass the project id to
the form to get the current project. CM and I discussed having a
different url and using the slug, but in the end I ended up with this
because its easiest and it worked. I'd be happy to look at different
ways of doing it though, I don't really like that different outputs
happen with the inclusion of a query param.

@ericronne, its difficult at this moment to get the "your donating"
part onto the same line as the 3 circles because the circles are shared
code with the rest of the site, so making that work would be a larger
architectural change. Also: the layout is off, the left margin doesn't
match whats below. We should talk about this because its due to grid
inconsistencies. It will be easy to fix once we have a grid.