Add icon font, font-awesome
Setup:
- includes font-awesome as an npm module.
- cps the sass files from font awesome into a convenient lib folder
under css src.
- copies the font-awesome fonts into a convenient fonts folder in root.
- modifies the font-awesome sass vars to point to the correct /fonts
folder. (this is the part I dislike the most, but sass will not compiled
correctly if I put the var in my code).
- load font-awesome css async through ajax

note: there's two ways to include font-awesome in this way: including
it in donation.css, the main css file, or including it as a separate
file, as I've done now. I picked my way because fonts probably aren't
required at start so shouldn't block critical rendering path. BUT,
further down the road, when we have a build task to remove all unused
css classes from the output, it may be better to move this to donation
css as it will be less file size in total. the biggest problem with
current way is stupid file-path var fix that I'm using in grunt.