Add a live style guide to static section
This is an auto generating css styleguide that can serve as a documented
reference and allow for a visual test environment.

It uses node-kss to find comments in the css files to generate the
guide.

There is both a guide and a template checked in. The template is written
while the actual guide is generated.