Simplified publication styleguide
Rather then messing with gh-pages, which would require a complicated
process to keep upstream and forks in sync, this serves the styleguide
with a service called raw git: rawgit.com.

This link will always point to the main site, not the fork, which I
think is fine because you can look at that one locally.