Share tool on Donate
Moved over all stylings, behavior and markup for the share tool.

Tested:
- Chrome
- Firefox
- Safari
- IE 11
- Iphone, both orientations
- IPad, both orientations

Scaled the beta panel down on mobile to make more room for the share too,
it also looks way better if it doens't drop to two lines.

Theres some overlap at a small amount of mobile sizes due to the page header
getting in the way. I think we should deal with this when we fix the header on
mobile.